Move closure string coercion into `impl Closure`
