--- crates/nu-protocol/src/engine/capture_block.rs
@@ -1,9 +1,0 @@
-use crate::{BlockId, Value, VarId};
-
-use serde::{Deserialize, Serialize};
-
-#[derive(Clone, Debug, Serialize, Deserialize)]
-pub struct Closure {
-    pub block_id: BlockId,
-    pub captures: Vec<(VarId, Value)>,
-}
