+++ crates/nu-protocol/src/engine/mod.rs
@@ -3,7 +3,7 @@ mod argument;
 mod cached_file;
 mod call;
 mod call_info;
-mod capture_block;
+mod closure;
 mod command;
 mod description;
 mod engine_state;
@@ -23,7 +23,7 @@ pub use cached_file::CachedFile;
 pub use argument::*;
 pub use call::*;
 pub use call_info::*;
-pub use capture_block::*;
+pub use closure::*;
 pub use command::*;
 pub use engine_state::*;
 pub use error_handler::*;
