+++ crates/nuon/src/to.rs
@@ -99,22 +99,7 @@ fn value_to_string(
         }
         Value::Closure { val, .. } => {
             if serialize_types {
-                let block = engine_state.get_block(val.block_id);
-                if let Some(span) = block.span {
-                    let contents_bytes = engine_state.get_span_contents(span);
-                    let contents_string = String::from_utf8_lossy(contents_bytes);
-                    Ok(contents_string.to_string())
-                } else {
-                    Err(ShellError::CantConvert {
-                        to_type: "string".into(),
-                        from_type: "closure".into(),
-                        span,
-                        help: Some(format!(
-                            "unable to retrieve block contents for closure with id {}",
-                            val.block_id.get()
-                        )),
-                    })
-                }
+                Ok(val.coerce_into_string(&engine_state, span)?.to_string())
             } else {
                 Err(ShellError::UnsupportedInput {
                     msg: "closures are currently not deserializable (use --serialize to serialize as a string)".into(),
